[Bloom filter]: Implementation near ended

+++ droxyd/src/bloom_filter/fill_filter.rs
@@ -1,7 +1,7 @@
-pub fn fill_filter(filter: &mut Vec<i32>, hash_funcs: Vec<fn(&str) -> i32>, word: &str) -> ()
+pub fn fill_filter(filter: &mut Vec<u8>, hash_funcs: Vec<fn(&str) -> i64>, word: &str) -> ()
 {
     for func in hash_funcs
     {
-        filter[func(word)] = 1;
+        filter[func(word) as usize] = 1;
     }
 }
